requirements.txt
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arro3-core==0.5.1
beautifulsoup4==4.13.4
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
deltalake==1.1.3
Deprecated==1.2.18
duckdb==1.3.2
idna==3.10
iniconfig==2.1.0
lxml==6.0.0
minio==7.2.16
packaging==25.0
pluggy==1.6.0
pycparser==2.22
pycryptodome==3.23.0
Pygments==2.19.2
pytest==8.4.1
PyYAML==6.0.2
requests==2.32.4
soupsieve==2.7
typing_extensions==4.14.1
urllib3==2.5.0
wrapt==1.17.2
